# swagger.yaml
openapi: 3.0.3
info:
  title: Helpp API
  description: |
    API do projeto **Helpp** — plataforma de ajuda mútua que conecta pessoas que precisam de ajuda com quem pode ajudar.

    ## Autenticação
    A maioria dos endpoints protegidos usa **Bearer Token (JWT)**. O access token expira em **15 minutos**.
    Use `POST /auth/refresh` para renová-lo com o refresh token (validade de 30 dias).

    ## Sistema de Reciprocidade
    Cada usuário possui um saldo de reciprocidade:
    - `helpReceived` incrementa quando recebe uma ajuda concluída
    - `helpGiven` incrementa quando conclui uma ajuda como helper
    - Se `helpReceived > helpGiven`, o usuário está bloqueado de criar novos pedidos até ajudar alguém
  version: 1.0.0
  contact:
    name: Helpp
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
            profileComplete:
              type: boolean
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        nickname:
          type: string
        cpfSuffix:
          type: string
          description: Últimos 3 dígitos do CPF
        phoneSuffix:
          type: string
          description: Últimos 4 dígitos do telefone
        city:
          type: string
        neighborhood:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        avatar:
          type: string
          description: Imagem em base64
        profileComplete:
          type: boolean
        helpReceived:
          type: integer
          description: Total de ajudas recebidas
          default: 0
        helpGiven:
          type: integer
          description: Total de ajudas dadas
          default: 0
    Request:
      type: object
      properties:
        _id:
          type: string
        userId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
        category:
          type: string
          enum:
            - Food or Groceries
            - Medicine
            - Transportation
            - Temporary Shelter
            - Help with Bills
            - Job Opportunity
            - School Supplies
            - Emotional Support
            - Other
        reason:
          type: string
        urgency:
          type: string
          enum: [Low, Medium, High, Emergency]
        giveBack:
          type: string
        status:
          type: string
          enum: [pending, matched, completed, cancelled]
          default: pending
        latitude:
          type: number
        longitude:
          type: number
        createdAt:
          type: string
          format: date-time
    Match:
      type: object
      properties:
        _id:
          type: string
        requestId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Request'
        helperId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
        offerMessage:
          type: string
        status:
          type: string
          enum: [pending, active, awaiting_confirmation, completed, cancelled]
          default: pending
        proofImage:
          type: string
          description: Imagem de comprovante em base64
        createdAt:
          type: string
          format: date-time
    Message:
      type: object
      properties:
        _id:
          type: string
        matchId:
          type: string
        senderId:
          type: object
          properties:
            _id:
              type: string
            name:
              type: string
            nickname:
              type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Cria nova conta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  example: João Silva
                email:
                  type: string
                  format: email
                  example: joao@email.com
                password:
                  type: string
                  minLength: 6
                  example: senha123
      responses:
        '201':
          description: Conta criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          description: Campos obrigatórios faltando
        '409':
          description: Email já em uso
  /auth/login:
    post:
      tags: [Auth]
      summary: Autentica e retorna tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Credenciais inválidas
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Renova o access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token renovado
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Refresh token inválido ou expirado
...
  /matches:
    get:
      tags: [Matches]
      summary: Lista matches com filtros
      parameters:
        - in: query
          name: helperId
          schema:
            type: string
        - in: query
          name: requestId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, active, awaiting_confirmation, completed, cancelled]
      responses:
        '200':
          description: Lista de matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'

    post:
      tags: [Matches]
      summary: Helper envia oferta de ajuda
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestId]
              properties:
                requestId:
                  type: string
                offerMessage:
                  type: string
      responses:
        '201':
          description: Oferta criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '409':
          description: Pedido já tem helper ativo ou oferta duplicada
  /matches/mine:
    get:
      tags: [Matches]
      summary: Lista matches do helper autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Matches do helper
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'
  /matches/{id}:
    patch:
      tags: [Matches]
      summary: Atualiza status do match
      description: |
        **Transições válidas:**
        - Receptor aceita oferta: `pending → active`
        - Receptor recusa oferta: `pending → cancelled`
        - Helper cancela match ativo: `active → cancelled`
        - Helper sinaliza conclusão: `active → awaiting_confirmation`
        - Helper reverte sinalização: `awaiting_confirmation → active`
        - Receptor confirma ajuda recebida: `active|awaiting_confirmation → completed` *(incrementa helpReceived/helpGiven)*
        - Helper marca diretamente como concluído: `active → completed` *(incrementa helpReceived/helpGiven)*
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, active, awaiting_confirmation, completed, cancelled]
                proofImage:
                  type: string
                  description: Base64 (opcional, usado na conclusão)
      responses:
        '200':
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          description: Transição de status inválida
        '403':
          description: Acesso negado
        '404':
          description: Match não encontrado
  /matches/{id}/proof:
    patch:
      tags: [Matches]
      summary: Atualiza foto de comprovante
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                proofImage:
                  type: string
                  nullable: true
                  description: Base64 da imagem ou null para remover
      responses:
        '200':
          description: Comprovante atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  proofImage:
                    type: string
                    nullable: true
  /chat/{matchId}:
    get:
      tags: [Chat]
      summary: Busca mensagens de um match
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: matchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mensagens do match em ordem cronológica
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '403':
          description: Usuário não é participante do match
        '404':
          description: Match não encontrado

    post:
      tags: [Chat]
      summary: Envia mensagem no chat do match
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: matchId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
      responses:
        '201':
          description: Mensagem enviada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Usuário não é participante do match
...